[#57]Refactor: new-portal folder 분리
