--- src/page/UploadPostNew.jsx
@@ -1,11 +1,0 @@
-import TextEditor from "../component/ new-portal/TextEditor";
-
-const UploadPostNew = () => {
-  return (
-    <div className="pl-[220px] pr-[30px] pt-[84px]">
-      <TextEditor />
-    </div>
-  );
-};
-
-export default UploadPostNew;
