+++ src/page/new-portal/UploadPost.jsx
@@ -1,0 +1,11 @@
+import TextEditor from "../../component/ new-portal/TextEditor";
+
+const UploadPost = () => {
+  return (
+    <div className="pl-[220px] pr-[30px] pt-[84px]">
+      <TextEditor />
+    </div>
+  );
+};
+
+export default UploadPost;
